refactor(compute): extract AgentService API to @dxos/compute
