+++ packages/core/compute/compute/package.json
@@ -18,6 +18,11 @@
       "types": "./dist/types/src/index.d.ts",
       "default": "./dist/lib/neutral/index.mjs"
     },
+    "./AgentService": {
+      "source": "./src/AgentService.ts",
+      "types": "./dist/types/src/AgentService.d.ts",
+      "default": "./dist/lib/neutral/AgentService.mjs"
+    },
     "./StorageService": {
       "source": "./src/StorageService.ts",
       "types": "./dist/types/src/StorageService.d.ts",
