+++ packages/core/compute/functions-runtime/src/agent-service/AgentService.ts
@@ -5,14 +5,13 @@
 // @import-as-namespace
 
 import * as Cause from 'effect/Cause';
-import * as Context from 'effect/Context';
 import * as Effect from 'effect/Effect';
 import * as Layer from 'effect/Layer';
-import * as Stream from 'effect/Stream';
 
-import { ModelName } from '@dxos/ai';
+import type { ModelName } from '@dxos/ai';
 import { AiContext } from '@dxos/assistant';
-import { type Trace, Blueprint, McpServer, Process } from '@dxos/compute';
+import { Blueprint, McpServer, Process } from '@dxos/compute';
+import { AgentService, type GetSessionOptions, getSession, type Service, type Session } from '@dxos/compute/AgentService';
 import { ProcessManager } from '@dxos/compute-runtime';
 import { Database, Feed, Obj, Ref, Registry } from '@dxos/echo';
 import { EffectEx } from '@dxos/effect';
@@ -23,80 +22,11 @@ import { AGENT_PROCESS_KEY, AgentProcess } from './agent-process';
 import { type CompletionGuard } from './completion-guard';
 import { type DelegationStrategy } from './delegation-strategy';
 
+// Re-export the pure API from @dxos/compute so callers using this namespace get the full surface.
+export { AgentService, type GetSessionOptions, getSession, hydrate, type Service, type Session } from '@dxos/compute/AgentService';
+
 const isTerminalProcess = (state: Process.State): boolean =>
   state === Process.State.SUCCEEDED || state === Process.State.FAILED || state === Process.State.TERMINATED;
-
-export interface Service {
-  /**
-   * Gets or creates a session for a feed.
-   */
-  getSession: (feed: Feed.Feed, options?: GetSessionOptions) => Effect.Effect<Session>;
-
-  /**
-   * Hydrates agent processes persisted by a previous session.
-   * Each record is rehydrated with a fresh {@link AgentProcess} built from the layer options.
-   */
-  hydrate: () => Effect.Effect<void>;
-}
-
-export class AgentService extends Context.Tag('@dxos/functions-runtime/AgentService')<AgentService, Service>() {}
-
-/**
- * Handle to an agent session.
- */
-export interface Session {
-  /**
-   * The feed that the session is associated with.
-   */
-  readonly feed: Feed.Feed;
-
-  /**
-   * Gets the context objects from the agent.
-   */
-  getContext: () => Effect.Effect<Ref.Ref<Obj.Unknown>[], never, Database.Service>;
-
-  /**
-   * Adds context objects to the agent.
-   */
-  addContext: (context: Ref.Ref<Obj.Unknown>[]) => Effect.Effect<void, never, Database.Service>;
-
-  /**
-   * Submits a prompt to the agent.
-   */
-  submitPrompt: (prompt: string) => Effect.Effect<void>;
-
-  /**
-   * Wait until agent has completed its work.
-   */
-  waitForCompletion: () => Effect.Effect<void>;
-
-  /**
-   * Terminates the agent process and clears its durable storage.
-   */
-  terminate: () => Effect.Effect<void>;
-
-  /**
-   * Subscribe to ephemeral trace events (e.g., streaming partial messages).
-   * Replays buffered events, then streams new ones until the process ends.
-   *
-   * When forking a collector from a short-lived parent (e.g. `useEffect` +
-   * `runPromise(Effect.forEach(subscribe))`), use {@link Effect.forkDaemon} so the
-   * stream survives after the parent scope closes; interrupt it on dispose.
-   */
-  subscribeEphemeral: () => Stream.Stream<Trace.Message>;
-}
-
-/**
- * Gets or creates a session for a feed.
- */
-export const getSession = Effect.serviceFunctionEffect(AgentService, (service) => service.getSession);
-
-export const hydrate = Effect.serviceFunctionEffect(AgentService, (service) => service.hydrate);
-
-export interface GetSessionOptions {
-  readonly model?: ModelName;
-  readonly systemPrompt?: string;
-}
 
 export interface CreateSessionOptions {
   readonly blueprints?: Blueprint.Blueprint[];
